requirements.txt
streamlit
pandas
numpy
matplotlib.pyplot
mysqlclient==2.1.1
SQLAlchemy==2.0.16